server/app/requirements.txt
flask==2.3.3
psycopg2-binary==2.9.9
flask-cors==3.0.10
